requirements.txt
librosa
matplotlib
numpy
soundfile
ffmpeg-python
tensorflow==2.12
pandas
numpy
scikit-learn
matplotlib
dotenv
psycopg2
gspread
oauth2client
psycopg2-binary
yt-dlp
streamlit
tensorflow
numpy
pillow
librosa
matplotlib

